pyproject.toml
[tool.poetry]
name = "pyproject-toml"
version = "0.1.0"
description = ""
authors = ["tjazerzen <tjaz.erzen@zerodays.dev>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
bcrypt = "^4.2.0"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
pyjwt = "^2.9.0"
psycopg2-binary = "^2.9.10"
openai = "^1.57.1"
python-multipart = "^0.0.19"
meilisearch = "^0.33.0"
strawberry-graphql = "^0.258.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \venv
  | \build
  | \dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src"]
